requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
pydantic
pydantic-settings
pillow
passlib[bcrypt]
